feat(log): implement log package for managing log segments
The log package provides functionality for managing log segments. It includes the following changes:

- Define Config struct to hold segment configuration
- Implement Log struct to manage multiple log segments
- Add NewLog function to create a new Log instance with given directory and configuration
- Implement setup method to initialize log segments from existing files or create a new one
- Add Append method to append a record to the active segment and create a new segment if the current one is maxed out
- Implement Read method to read a record at a given offset from the appropriate segment
- Add newSegment method to create a new segment with the given base offset
- Implement Close, Remove, and Reset methods for cleaning up log segments
- Add LowestOffset and HighestOffset methods to return the lowest and highest offsets across all segments
- Implement Truncate method to remove segments whose highest offset is lower than a given offset
- Add Reader method to return a reader that reads from all segments

The log package provides a way to store and retrieve records in a log structure divided into segments. It allows appending records, reading records at specific offsets, and managing the lifecycle of log segments.

Test cases are added to verify the functionality of the log package, including appending and reading records, handling out-of-range offsets, initializing with existing segments, testing the reader, and truncating segments.

+++ internal/log/config.go
@@ -4,6 +4,6 @@ type Config struct {
 	Segment struct {
 		MaxStoreBytes uint64
 		MaxIndexBytes uint64
-		InitialOffset uint32
+		InitialOffset uint64
 	}
 }
